+++ CHANGELOG.md
@@ -13,6 +13,7 @@ PowerModels.jl Change Log
 - Refactor `test/runtests.jl` (#962) (#965)
 - Remove empty docstrings (#963) (#966)
 - Add PrecompileTools (#967)
+- Add tests for `calc_admittance_matrix_inv` (#970)
 
 ### v0.21.3
 - Fix no-buses bug in `calc_connected_components` (#933)